+++ R-package/R/xgb.dump.R
@@ -45,10 +45,10 @@ xgb.dump <- function(model = NULL, fname = NULL, fmap = "", with_stats=FALSE, ..
   model_dump <- .Call("XGBoosterDumpModel_R", model$handle, fmap, as.integer(with_stats), PACKAGE = "xgboost")
 
   if (is.null(fname)) 
-    model_dump <- str_replace_all(model_dump, '\t', '')
+    model_dump <- stri_replace_all_regex(model_dump, '\t', '')
   
-  model_dump <- unlist(str_split(model_dump, '\n'))
-  model_dump <- grep('(^$|^0$)', model_dump, invert = TRUE, value = TRUE)
+  model_dump <- unlist(stri_split_regex(model_dump, '\n'))
+  model_dump <- grep('^\\s*$', model_dump, invert = TRUE, value = TRUE)
   
   if (is.null(fname)) {
     return(model_dump)