+++ jvm-packages/xgboost4j-spark/src/main/scala/ml/dmlc/xgboost4j/scala/spark/XGBoostModel.scala
@@ -20,15 +20,16 @@ import org.apache.hadoop.fs.{Path, FileSystem}
 import org.apache.spark.{TaskContext, SparkContext}
 import org.apache.spark.mllib.linalg.{DenseVector, Vector}
 import org.apache.spark.rdd.RDD
-import ml.dmlc.xgboost4j.java.{DMatrix => JDMatrix}
-import ml.dmlc.xgboost4j.scala.{DMatrix, Booster}
+import ml.dmlc.xgboost4j.java.{Rabit, DMatrix => JDMatrix}
+import ml.dmlc.xgboost4j.scala.{Booster, DMatrix}
+import scala.collection.JavaConverters._
 
 class XGBoostModel(_booster: Booster) extends Serializable {
 
   /**
-   * Predict result with the given testset (represented as RDD)
+   * Predict result with the given test set (represented as RDD)
    *
-   * @param testSet test set representd as RDD
+   * @param testSet test set represented as RDD
    * @param useExternalCache whether to use external cache for the test set
    */
   def predict(testSet: RDD[Vector], useExternalCache: Boolean = false): RDD[Array[Array[Float]]] = {
@@ -37,6 +38,8 @@ class XGBoostModel(_booster: Booster) extends Serializable {
     val appName = testSet.context.appName
     testSet.mapPartitions { testSamples =>
       if (testSamples.hasNext) {
+        val rabitEnv = Array("DMLC_TASK_ID" -> TaskContext.getPartitionId().toString).toMap
+        Rabit.init(rabitEnv.asJava)
         val cacheFileName = {
           if (useExternalCache) {
             s"$appName-dtest_cache-${TaskContext.getPartitionId()}"
@@ -45,7 +48,9 @@ class XGBoostModel(_booster: Booster) extends Serializable {
           }
         }
         val dMatrix = new DMatrix(new JDMatrix(testSamples, cacheFileName))
-        Iterator(broadcastBooster.value.predict(dMatrix))
+        val res = broadcastBooster.value.predict(dMatrix)
+        Rabit.shutdown()
+        Iterator(res)
       } else {
         Iterator()
       }
@@ -53,8 +58,9 @@ class XGBoostModel(_booster: Booster) extends Serializable {
   }
 
   /**
-   * Predict result with the given testset (represented as RDD)
-   * @param testSet test set representd as RDD
+   * Predict result with the given test set (represented as RDD)
+   *
+   * @param testSet test set represented as RDD
    * @param missingValue the specified value to represent the missing value
    */
   def predict(testSet: RDD[DenseVector], missingValue: Float): RDD[Array[Array[Float]]] = {
@@ -78,10 +84,39 @@ class XGBoostModel(_booster: Booster) extends Serializable {
   }
 
   /**
-   * predict result given the test data (represented as DMatrix)
+   * Predict result with the given test set (represented as DMatrix)
+   *
+   * @param testSet test set represented as DMatrix
    */
   def predict(testSet: DMatrix): Array[Array[Float]] = {
     _booster.predict(testSet, true, 0)
+  }
+
+  /**
+   * Predict leaf instances with the given test set (represented as RDD)
+   *
+   * @param testSet test set represented as RDD
+   */
+  def predictLeaves(testSet: RDD[Vector]): RDD[Array[Array[Float]]] = {
+    import DataUtils._
+    val broadcastBooster = testSet.sparkContext.broadcast(_booster)
+    testSet.mapPartitions { testSamples =>
+      if (testSamples.hasNext) {
+        val dMatrix = new DMatrix(new JDMatrix(testSamples, null))
+        Iterator(broadcastBooster.value.predictLeaf(dMatrix, 0))
+      } else {
+        Iterator()
+      }
+    }
+  }
+
+  /**
+   * Predict leaf instances with the given test set (represented as DMatrix)
+   *
+   * @param testSet test set represented as DMatrix
+   */
+  def predictLeaves(testSet: DMatrix): Array[Array[Float]] = {
+    _booster.predictLeaf(testSet, 0)
   }
 
   /**
@@ -97,7 +132,7 @@ class XGBoostModel(_booster: Booster) extends Serializable {
   }
 
   /**
-   * get the booster instance of this model
+   * Get the booster instance of this model
    */
   def booster: Booster = _booster
 }